+++ src/js/settings-page.js
@@ -16,8 +16,8 @@ chrome.runtime.getBackgroundPage((page) => {
   const app = page.app;
 
   /*
-     Allows user to switch between capturing a video at its natural size or
-     at its visible size.
+     Control the size of a video screenshot.
+     Possible values: 'visible', 'natural'.
   */
   const vSizeSelect = document.getElementById('video-size');
   setDefaultOption(vSizeSelect, app.settings.getItem('videoSize'));
@@ -27,8 +27,8 @@ chrome.runtime.getBackgroundPage((page) => {
   });
 
   /*
-    Allows  user to select the number of screenshots stored in temporary
-    browser memory.
+    Control the number of screenshots stored in memory.
+    Possible values: 4, 8, 16, 32, 64, 99, Infinity
   */
   const maxSelect = document.getElementById('max-screenshots');
   setDefaultOption(maxSelect, app.settings.getItem('maxScreenshots'));
@@ -45,11 +45,22 @@ chrome.runtime.getBackgroundPage((page) => {
     }
   });
 
+  /*
+    Control whether or not tooltips are shown on the settings and help icons.
+    Possible values: 1 (show), 0 (hide)
+  */
+  const tooltipSelect = document.getElementById('show-tooltips');
+  setDefaultOption(tooltipSelect, app.settings.getItem('showTooltips'));
+  tooltipSelect.addEventListener('change', (event) => {
+    const option = event.target;
+    app.settings.setItem('showTooltips', Number(option.value));
+  });
+
   app.getKeyboardCommands().then((commands) => {
     includeHTML().then((nodes) => {
       commands.forEach(drawCommandHelp);
       feather.replace();
-      tippyFooter();
+      tippyFooter(app);
       tippyHelp();
     });
   });